Merging from Addfeature branch where all the enhancements were added and errors fixed

+++ ImageHoster/src/main/java/ImageHoster/controller/ImageController.java
@@ -45,11 +45,12 @@ public class ImageController {
     //Also now you need to add the tags of an image in the Model type object
     //Here a list of tags is added in the Model type object
     //this list is then sent to 'images/image.html' file and the tags are displayed
-    @RequestMapping("/images/{title}")
-    public String showImage(@PathVariable("title") String title, Model model) {
-        Image image = imageService.getImageByTitle(title);
+    @RequestMapping("/images/{imageId}/{title}")
+    public String showImage(@PathVariable("imageId") int id, Model model) {
+        Image image = imageService.getImage(id);
         model.addAttribute("image", image);
         model.addAttribute("tags", image.getTags());
+        model.addAttribute("comments", image.getComments());  //to display all the comments related to the image selected
         return "images/image";
     }
 
@@ -92,13 +93,26 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
-
+        User currentUser = (User) session.getAttribute("loggeduser");
+        User imgOwnerUser = image.getUser();
+        boolean authorizedUser = checkUserAuthorization(currentUser, imgOwnerUser);
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+
+        if(authorizedUser) {
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        }
+        else {
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("comments", image.getComments());
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("editError", error);
+            return "images/image";
+        }
+
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -132,7 +146,7 @@ public class ImageController {
         updatedImage.setDate(new Date());
 
         imageService.updateImage(updatedImage);
-        return "redirect:/images/" + updatedImage.getTitle();
+        return "redirect:/images/" + updatedImage.getId() + "/" +updatedImage.getTitle();
     }
 
 
@@ -140,9 +154,27 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model,HttpSession session) {
+        Image image = imageService.getImage(imageId);
+        User currentUser = (User) session.getAttribute("loggeduser");
+        User imgOwnerUser = image.getUser();
+        boolean authorizedUser = checkUserAuthorization(currentUser, imgOwnerUser);
+        if(authorizedUser) {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
+        else {
+            String error = "Only the owner of the image can delete the image";
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("comments", image.getComments());
+            model.addAttribute("deleteError", error);
+            return "images/image";
+        }
+
+
+
     }
 
 
@@ -182,9 +214,24 @@ public class ImageController {
             tagString.append(tags.get(i).getName()).append(",");
         }
 
-        Tag lastTag = tags.get(tags.size() - 1);
-        tagString.append(lastTag.getName());
 
+        if(tags.size() == 0) {
+            tagString = tagString.append("");
+        }
+        else {
+            Tag lastTag = tags.get(tags.size() - 1);
+            tagString.append(lastTag.getName());
+        }
         return tagString.toString();
+    }
+
+    //This method checks whether the current user is the owner of the image he is trying to delete or edit
+    private boolean checkUserAuthorization(User currUser, User imgOwn) {
+        if(currUser.getId() == imgOwn.getId()) {
+            return true;
+        }
+        else {
+            return false;
+        }
     }
 }
